feat(canary-web-components): allow specifying target for anchor in ic-data-table cell
Allow specifying target attribute for the anchor tag in the cell, along with the href. When the
target is "_blank", then rel="noopener noreferrer" is also automatically added anchor tag. There are
existing styles that add a new window icon to the link when the target is "_blank" and so no change
is required to achieve that.

Closes: #2751